just-in-time-provisioning/iot_client/requirements.txt
awscrt==0.17.0
awsiotsdk==1.17.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
idna==3.4
pycparser==2.21
pyOpenSSL==23.2.0
requests==2.31.0
retrying==1.3.4
six==1.16.0
urllib3==2.0.4
